Aplicación de OclusionCulling en el juego

+++ Espresso_VR/Assets/Scripts/PlayerMove.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerMove : MonoBehaviour
 {
-    public int playerSpeed;
+    public float playerSpeed;
     public Transform vrCamera;
     public float toggleAngle;
